move create organization test util to enterprise
